+++ src/Rules/Date.php
@@ -11,23 +11,30 @@ declare(strict_types = 1);
 
 namespace Linna\Filter\Rules;
 
+use Linna\Filter\AbstractDate;
 use DateTime;
 
 /**
  * Check if one date is valid.
  */
-class Date
+class Date extends AbstractDate
 {
     /**
      * @var array Arguments expected.
      */
     private $arguments = ['string'];
-    
+
     /**
-     * @var DateTime Valid date.
+     * @var string Valid date.
      */
     private $date;
-    
+
+    /**
+     *
+     * @var DateTime Valid date in DateTime object.
+     */
+    private $dateTimeObject;
+
     /**
      * Validate.
      *
@@ -35,8 +42,35 @@ class Date
      */
     public function validate($received, string $format): bool
     {
-        $date = date_parse_from_format($format, $received);
+        if ($this->parseDate($received, $format)) {
+            return true;
+        }
         
+        $this->date = $received;
+        
+        $dateReceived = DateTime::createFromFormat($format, $received);
+        
+        if ($this->dateHaveNoTime($format)) {
+            $dateReceived->setTime(0, 0, 0);
+        }
+        
+        $this->dateTimeObject = $dateReceived;
+        
+        return false;
+    }
+
+    /**
+     * Parse date.
+     *
+     * @param type $received
+     * @param string $format
+     *
+     * @return bool
+     */
+    private function parseDate($received, string $format) : bool
+    {
+        $date = date_parse_from_format($format, $received);
+
         //set to zero the date
         $month = $day = null;
         
@@ -47,54 +81,29 @@ class Date
         
         settype($month, 'bool');
         settype($day, 'bool');
-        
-        
-        
+
         if (!$month) {
             return true;
         }
-        
+
         if (!$day) {
             return true;
         }
-        
+
         if ($warning_count + $error_count) {
             return true;
         }
-
-        $this->date = $date;
-
+        
         return false;
     }
     
     /**
-     * Sanitize.
+     * Return DateTime object.
      *
-     * @param mixed $value
+     * @return DateTime
      */
-    public function sanitize(&$value): void
+    public function getDateTimeObject(): DateTime
     {
-        //set to zero the date
-        $year = $month = $day = null;
-        
-        //set to zero the time
-        $hour = $minute = $second = $fraction = null;
-        
-        extract($this->date, EXTR_IF_EXISTS);
-
-        //force type if there is bool value
-        settype($year, 'integer');
-        settype($month, 'integer');
-        settype($day, 'integer');
-        
-        settype($hour, 'integer');
-        settype($minute, 'integer');
-        settype($second, 'integer');
-        
-        settype($fraction, 'float');
-        
-        $second += $fraction;
-        
-        $value = new DateTime("{$year}-{$month}-{$day} {$hour}:{$minute}:{$second}");
+        return $this->dateTimeObject;
     }
 }